strvec: implement swapping two strvecs
In a future commit, we'll want the ability to efficiently swap the
contents of two strvec instances without needing to copy any data.
Since a strvec is simply a pointer and two sizes, swapping them is as
simply as copying the two pointers and sizes, so let's do that.

We use a temporary here simply because C doesn't provide a standard
swapping function, unlike C++ and Rust, but a good optimizing compiler
will recognize this syntax and handle it appropriately using an
optimization pass.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>